requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
fastapi>=0.68.0,<0.69.0
httpx>=0.24.0,<0.25.0
idna==3.10
jiter==0.10.0
openai==1.82.0
pydantic>=1.8.0,<2.0.0
pydantic_core==2.33.2
python-dotenv>=0.19.0,<0.20.0
sniffio==1.3.1
starlette==0.14.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions>=4.0.0,<5.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
email-validator>=1.1.3,<1.2.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.7.0,<1.8.0
psycopg2-binary>=2.9.0,<3.0.0
redis>=4.0.0,<4.1.0
aiohttp>=3.8.0,<3.9.0
python-socketio>=5.0.0,<5.1.0
websockets>=10.0,<11.0
pytest>=7.0.0,<7.1.0
pytest-asyncio>=0.18.0,<0.19.0
pytest-cov>=3.0.0,<3.1.0
black>=23.1.0,<24.0.0
isort>=5.10.0,<5.11.0
flake8>=4.0.0,<4.1.0
mypy>=0.910,<0.911
